+++ src/RequestParser/RequestParser.cpp
@@ -1,8 +1,5 @@
 #include "RequestParser.hpp"
-#include "../utils/utils.hpp"
 #include "../GlobalConfig/GlobalConfig.hpp"
-#include "../ClientSession/ClientSession.hpp"
-#include <stdexcept>
 #include <sstream>
 
 //Body크기에 제한 없이 테스트하고 싶은 경우, BODY_MAX_LENGTH대신 std::numeric_limits을 사용할 수 있음
@@ -119,250 +116,4 @@ EnumReqStatus RequestParser::handleCRLFLine(const EnumReqStatus &curStatus) {
 		return REQ_HEADER_CRLF;// MUST TO DO: 이 상태에서 request가 새롭게 들어오면 DONE을 해줘야하는 상황이 있음
 	return REQ_ERROR;
 	// MUST TO DO: 이상태에서 DONE의 판단을 할 수 있는가?
-}
-
-// 2. Start-line 한 줄을 파싱하는 함수
-// line: \r\n기준으로 나뉜 요청 데이터의 첫 줄
-// reqMsg: 현재 요청 데이터를 파싱하고 저장할 RequestMessage
-EnumStatusCode RequestParser::parseStartLine(const std::string &line, RequestMessage &reqMsg) {
-	std::istringstream iss(line);
-	std::string buffer;
-
-	if (std::getline(iss, buffer, ' ')) {
-		if (buffer == "GET")
-			reqMsg.setMethod(GET);
-		else if (buffer == "POST")
-			reqMsg.setMethod(POST);
-		else if (buffer == "DELETE")
-			reqMsg.setMethod(DELETE);
-		else {
-			reqMsg.setStatus(REQ_ERROR);
-			return NOT_IMPLEMENTED;//status code: 구현되지 않은 메서드
-		}
-	} else {
-		return BAD_REQUEST;
-	}
-
-	if (std::getline(iss, buffer, ' ')) {
-		if (buffer.size() > this->uriMaxLength_)
-			return URI_TOO_LONG;//status code: URI가 서버지원사이즈보다 큼
-		reqMsg.setTargetURI(buffer);
-	} else {
-		return BAD_REQUEST;//status code: start-line의 "method URI version" 형식이 유효하지 않음
-	}
-
-	iss >> buffer;
-	if (buffer == "HTTP/1.1")
-		reqMsg.setStatus(REQ_STARTLINE);
-	else {
-		reqMsg.setStatus(REQ_ERROR);
-		return BAD_REQUEST;//status code: HTTP 버전 오류
-	}
-	return NONE_STATUS_CODE;
-}
-
-// 3. field-line 한 줄을 파싱하는 함수
-// line: \r\n기준으로 나뉜 요청 데이터, 첫 줄(start line)이 아닌 다음 줄
-// reqMsg: 현재 요청 데이터를 파싱하고 저장할 RequestMessage
-EnumStatusCode RequestParser::parseFieldLine(const std::string &line, RequestMessage &reqMsg) {
-	std::istringstream iss(line);
-	std::string name;
-	std::string value;
-	std::vector<std::string> values;
-
-	// 3-1. field-line의 기본 파싱 (name: value, ...)
-	std::getline(iss, name, ':');
-	if (name == "Host" || name == "Content-Length"
-	||  name == "Connection" || name == "Transfer-Encoding"
-	||  name == "User-Agent" ||  name == "Authorization"
-	||  name == "Referer" ||  name == "Range"
-	||  name == "If-Modified-Since" ||  name == "If-Unmodified-Since"
-	||  name == "If-Match" ||  name == "If-None-Match"
-	||  name == "Content-Location") {
-		std::getline(iss, value, '\0');
-		values.push_back(utils::strtrim(value));
-	} else {
-		while (std::getline(iss, value, ',')) {
-			value = utils::strtrim(value);
-			values.push_back(value);
-		}
-	}
-	// map<string, vector> 형태의 멤버변수 데이터에 저장
-	reqMsg.addFieldLine(name, values);
-	
-	// 3-2. field value 갯수 검증
-	if (!validateFieldValueCount(name, values.size()))
-		return BAD_REQUEST;
-		
-	// 3-3. field value중 RequestMessage의 메타데이터 처리
-	if (!handleFieldValue(name, *values.begin(), reqMsg))
-		return BAD_REQUEST;
-	
-	// 3-4. RequestMessage가 하나 이상의 field-line를 갖고 있으며, 아직 CRLF가 나오지 않음을 뜻함
-	reqMsg.setStatus(REQ_HEADER_FIELD);
-	return NONE_STATUS_CODE;
-}
-
-// 3-2. field-line의 value 갯수를 검증하는 함수
-bool RequestParser::validateFieldValueCount(const std::string &name, const int count) {
-	if (count < 1){
-		return false;
-	}
-	(void)name;
-	//if (count > 1
-	//&& (name == "Host" || name == "Content-Length"
-	//||  name == "Connection" || name == "Transfer-Encoding"
-	//||  name == "User-Agent" ||  name == "Authorization"
-	//||  name == "Referer" ||  name == "Range"
-	//||  name == "If-Modified-Since" ||  name == "If-Unmodified-Since"
-	//||  name == "If-Match" ||  name == "If-None-Match"
-	//||  name == "Content-Location")){
-	//	return false;
-	//}
-	return true;
-}
-
-// 3-3. field-line의 value중 RequestMessage의 메타데이터에 해당하는 value 파싱하는 함수
-bool RequestParser::handleFieldValue(const std::string &name, const std::string &value, RequestMessage &reqMsg) {
-	if (name == "Host") {//									1) Host
-		reqMsg.setMetaHost(value);
-	} else if (name == "Connection") {//					2) Connection
-		EnumConnect connection;
-		if (value == "keep-alive")
-			connection = KEEP_ALIVE;
-		else if (value == "close")
-			connection = CLOSE;
-		else
-			return false;
-		reqMsg.setMetaConnection(connection);
-	} else if (name == "Content-Length") {//				3) Content-Length
-		size_t contentLength;
-		try {
-			contentLength = utils::sto_size_t(value);
-		} catch (const std::exception & e) {
-			return false;
-		}
-		reqMsg.setMetaContentLength(contentLength);
-	} else if (name == "Transfer-Encoding") {//				4) Transfer-Encoding
-		EnumTransEnc transferEncoding;
-		if (value == "chunked")
-			transferEncoding = CHUNK;
-		else
-			return false;
-		reqMsg.setMetaTransferEncoding(transferEncoding);
-	} else if (name == "Content-Type") {//					5) Content-Type
-		reqMsg.setMetaContentType(value);
-	}
-	return true;
-}
-
-// 4. Body를 파싱하는 함수, Content-Length의 길이에 따라 예외 처리됨. 유효한 문자열은 기존의 Body의 추가함
-// readBuffer: Body로 파싱할 요청 데이터이자, 남은 데이터를 저장할 ClientSession의 readBuffer
-// reqMsg: 현재 요청 데이터를 파싱하고 저장할 RequestMessage
-EnumStatusCode RequestParser::parseBody(std::string &readBuffer, RequestMessage &reqMsg) {
-	const size_t contentLength = reqMsg.getMetaContentLength();
-
-	// 4-1. 최종 body의 길이 >= 현재 저장된 body길이 + 파싱하려는 길이
-	if (contentLength >= reqMsg.getBodyLength() + readBuffer.length()) {
-		if (this->bodyMaxLength_ < reqMsg.getBodyLength() + readBuffer.length())
-			return CONTENT_TOO_LARGE;//status code: Body가 설정보다 큼 “Request Entity Too Large”
-
-		reqMsg.addBody(readBuffer);
-		readBuffer = "";
-		if (contentLength == reqMsg.getBodyLength()) {
-			reqMsg.setStatus(REQ_DONE);
-			return NONE_STATUS_CODE;
-		} else {
-			reqMsg.setStatus(REQ_BODY);
-			return NONE_STATUS_CODE;
-		}
-	}
-
-	// 4-2. 최종 body의 길이 < 현재 저장된 body길이 + 파싱하려는 길이
-	// 우선, content length를 저장하되
-	size_t substrSize = contentLength - reqMsg.getBodyLength();
-	reqMsg.addBody(readBuffer.substr(0, substrSize));
-	std::string remainData = readBuffer.substr(substrSize);
-	
-	// 일차적으로 완성된 request이지만, 두 가지 우려사항을 검증해야함
-	const std::string methods[3] = {"GET", "POST", "DELETE"};
-	bool suspicious = true;
-	// 1) 남은 데이터가 유효한 메서드로 시작하면 -> 정상 요청 가능성
-	for (size_t i = 0; i < 3; ++i) {
-		if (remainData[0] == methods[i][0]) {
-			suspicious = false;
-			break;
-		}
-	}
-	// 2) body 안에 메서드가 들어있으면 → 의심스러운 요청
-	for (size_t i = 0; i < 3; ++i) {
-		if (reqMsg.getBody().find(methods[i]) != std::string::npos) {
-			suspicious = true;
-			break;
-		}
-	}
-
-	if (suspicious) {
-		reqMsg.setStatus(REQ_ERROR);
-		return BAD_REQUEST;//status code: 의심스러운 Body 거부
-	}
-	reqMsg.setStatus(REQ_DONE);
-	readBuffer = remainData;
-	return NONE_STATUS_CODE;
-}
-
-// 5. 청크 전송 파싱하는 함수. Transfer-Encoding이 chunked여서 Body가 청크로 아루어진 경우
-// readBuffer: Body로 파싱할 요청 데이터이자, 남은 데이터를 저장할 ClientSession의 readBuffer
-// reqMsg: 현재 요청 데이터를 파싱하고 저장할 RequestMessage
-EnumStatusCode RequestParser::cleanUpChunkedBody(std::string &readBuffer, RequestMessage &reqMsg) {
-	std::istringstream iss(readBuffer);
-	std::string buffer;
-	std::string tmp;
-	size_t chunkSize;
-
-	while (std::getline(iss, buffer, '\n')) {
-		if (iss.eof())// 5-1. \n으로 getline되지 않은 경우 readBuffer에 남겨 다음 loop에 진행
-			return NONE_STATUS_CODE;
-		if (!buffer.empty() && buffer[buffer.size() - 1] == '\r') {// 5-2. \r\n으로 찾은 줄 파싱
-			// iss에 남은 길이 측정을 위한 도구
-			std::streampos currentPos = iss.tellg();
-			iss.seekg(0, std::ios::end);
-			std::streampos endPos = iss.tellg();
-			iss.seekg(currentPos);
-
-			// 청크 사이즈 파싱
-			chunkSize = utils::sto_size_t(buffer.erase(buffer.size() - 1));
-			if (chunkSize == 0) {
-				reqMsg.setStatus(REQ_DONE);
-				reqMsg.setMetaContentLength(reqMsg.getBodyLength());
-				readBuffer = iss.str().substr(static_cast<std::string::size_type>(currentPos)); 
-				return NONE_STATUS_CODE;
-			}
-			
-			// 읽어낸 데이터가 모자를 경우 readBuffer에 돌려놓기 위함
-			tmp = buffer + "\r\n";
-			if (static_cast<size_t>(endPos - currentPos) < chunkSize + 2) {
-				readBuffer = tmp + iss.str().substr(static_cast<std::string::size_type>(currentPos));
-				return NONE_STATUS_CODE;
-			}
-
-			// 파싱한 청크사이즈 기준으로 청크 데이터를 읽음
-			buffer.resize(chunkSize + 2);
-			iss.read(&buffer[0], chunkSize + 2);
-
-			// 청크 데이터가 \r\n형식에 맞춰 들어왔는지 검증
-			if (buffer.compare(buffer.size() - 2, 2, "\r\n") != 0)
-				return BAD_REQUEST;//status code: 유효하지 않은 청크 데이터 형식
-				
-			if (this->bodyMaxLength_ < reqMsg.getBodyLength() + chunkSize)
-				return CONTENT_TOO_LARGE;//status code: Body가 설정보다 큼 “Request Entity Too Large”
-
-				// 청크 데이터에 \r\n를 제거하여 Body에 저장
-			reqMsg.addBody(buffer.substr(0, buffer.size() - 2));
-		} else { //\r\n으로 이루어져있지 않음
-			return BAD_REQUEST;//status code: CRLF가 아닌, 단일 LF
-		}
-	}
-	readBuffer = "";
-	return NONE_STATUS_CODE;
 }
